refactor: MemberCreateRq dto 클래스 삭제 후 record로 변경
