+++ backend/src/main/java/io/toytech/backend/domain/member/service/MemberService.java
@@ -1,7 +1,7 @@
 package io.toytech.backend.domain.member.service;
 
 import io.toytech.backend.domain.member.domain.Member;
-import io.toytech.backend.domain.member.dto.MemberCreateRq;
+import io.toytech.backend.domain.member.dto.MemberCreateRequest;
 import io.toytech.backend.domain.member.dto.MemberRs;
 import java.util.List;
 import java.util.Optional;
@@ -16,5 +16,5 @@ public interface MemberService {
 
   void delete(Long id);
 
-  MemberRs createDto(MemberCreateRq request);
+  MemberRs createDto(MemberCreateRequest request);
 }
